version 3

+++ script.sh
@@ -1,16 +1,24 @@
 #!/bin/bash
 if [ -e /tmp/out/lock ]
 then
-echo 'code erreur 22: script en cours d execution'
+    echo 'code erreur 22: script en cours d execution'
 else
     if [ -d /tmp/in ]
     then
-    touch /tmp/out/lock
-    echo 'OK!'
-    cp /tmp/in/* /tmp/out
-    chmod 775 /tmp/out
-    gzip /tmp/out/*
-    rm /tmp/out/lock
+        touch /tmp/out/lock
+        for file in /tmp/in/*
+        do
+            if [ -e $file ] && [ -f $file ]
+            then
+                echo 'OK!'
+                cp /tmp/in/* /tmp/out
+                chmod 775 /tmp/out
+                gzip /tmp/out/*
+                rm /tmp/out/lock.gz
+            else
+                echo  'code erreur 12: impossible d effectuer l operation'
+            fi
+        done
     else
     exit 2
     fi
